+++ packages/mapo-webapp/src/app/components/toolbar/toolbar.component.html
@@ -34,4 +34,13 @@
     <i class="ri-expand-right-fill"></i>
     <span class="subtext">e</span>
   </div>
+  <!-- <div
+    data-testid="toolbar-container"
+    class="tool create-container"
+    [ngClass]="{ enabled: tool === 'create-container' }"
+    (click)="selectTool('create-container')"
+  >
+    <i class="ri-checkbox-blank-line"></i>
+    <span class="subtext">b</span>
+  </div> -->
 </div>
