+++ packages/mapo-webapp/src/app/controllers/key-press.controller.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@angular/core';
 import { fabric } from 'fabric';
-import { EdgeStore } from '../../store/edge.store';
-import { TextNodeStore } from '../../store/text-node.store';
-import { Tool, ToolbarStore } from '../../store/toolbar.store';
-import { CanvasService } from '../canvas/canvas.service';
-import { ToolbarService } from './toolbar.service';
+import { EdgeStore } from '../store/edge.store';
+import { TextNodeStore } from '../store/text-node.store';
+import { Tool, ToolbarStore } from '../store/toolbar.store';
+import { CanvasService } from '../services/canvas/canvas.service';
+import { ToolbarService } from '../services/toolbar/toolbar.service';
 
 /**
  * When the canvas is active, listens for keyboard events, and controls the toolbar based on the actions pressed
@@ -12,7 +12,7 @@ import { ToolbarService } from './toolbar.service';
 @Injectable({
   providedIn: 'root',
 })
-export class ToolbarController {
+export class KeyPressController {
   canvas: fabric.Canvas | null = null;
 
   constructor(
@@ -57,6 +57,11 @@ export class ToolbarController {
 
     if (e.key === 'e') {
       this.toolbarService.selectOrCancel(Tool.CREATE_EDGE);
+      return;
+    }
+
+    if (e.key === 'b') {
+      this.toolbarService.selectOrCancel(Tool.CREATE_CONTAINER);
       return;
     }
