+++ packages/mapo-webapp/src/app/controllers/textnode-options.controller.ts
@@ -1,12 +1,16 @@
 import { fabric } from "fabric";
 import { Injectable } from "@angular/core";
-import { TextNodeOptionsStore } from "../../store/textnode-options.store";
-import { CanvasService } from "../canvas/canvas.service";
-import { TextNodeStore } from "../../store/text-node.store";
-import { SelectionService } from "../selection/selection.service";
-import { ToolbarStore } from "../../store/toolbar.store";
-import { BottomToolbarStore } from "../../store/bottom-toolbar.store";
+import { TextNodeOptionsStore } from "../store/textnode-options.store";
+import { CanvasService } from "../services/canvas/canvas.service";
+import { TextNodeStore } from "../store/text-node.store";
+import { SelectionController } from "./selection.controller";
+import { BottomToolbarStore } from "../store/bottom-toolbar.store";
 
+/**
+ * A controller listening to color changes and selection changes,
+ * for updating nodes when new colors are selected, and updating the colorbox
+ * when new nodes are selected.
+ */
 @Injectable({
   providedIn: 'root',
 })
@@ -18,7 +22,7 @@ export class TextNodeOptionsController {
     private textNodeOptionsStore: TextNodeOptionsStore,
     private canvasService: CanvasService,
     private textNodeStore: TextNodeStore,
-    private selectionService: SelectionService,
+    private selectionService: SelectionController,
     private bottomToolbarStore: BottomToolbarStore,
   ) {
     this.onBootstrap();
