+++ packages/mapo-webapp/src/app/services/clipboard/clipboard.service.ts
@@ -5,7 +5,7 @@ import { Edge } from "../../models/edge.model";
 import { TextNode } from "../../models/textnode.model";
 import { EdgeStore } from "../../store/edge.store";
 import { TextNodeStore } from "../../store/text-node.store";
-import { ClipboardData} from "./clipboard-data.model";
+import { ClipboardData} from "../../models/clipboard-data.model";
 import { FabricUtils } from "../../utils/fabric-utils";
 
 @Injectable({
