+++ packages/mapo-webapp/src/app/store/toolbar.store.ts
@@ -7,6 +7,7 @@ export enum Tool {
   CREATE_TEXT_NODE = 'create-text-node',
   CREATE_COMMENT_NODE = 'create-comment-node',
   CREATE_EDGE = 'create-edge',
+  CREATE_CONTAINER = 'create-container', 
   EDIT_TEXT_NODE = 'edit-text-node', // Not a real tool, but a valid state the user can be in
 }
